requirements.txt
streamlit
openai-whisper
pydub
deep-translator
httpx
numpy
noisereduce